requirements/base.txt
# Python Version 3.11
# ----------------------------------------------
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2
django-cors-headers==4.0.0
django-environ==0.10.0
drf-yasg==1.21.5
pyotp==2.8.0
Pillow==9.5.0



# Additional Packages
# -------------------------------------------
google-cloud-tasks==2.5.1  # https://github.com/googleapis/python-tasks
xrpl-py==3.0.0
cryptoconditions==0.8.1
fcm-django==2.2.1
django-money==3.5.3
certifi==2024.7.4
